+++ src/Main.java
@@ -6,7 +6,7 @@ public class Main {
         double height = 1.87; //Рост
         int weight = 98; //Вес
 
-        double index = (int) service.calculate(height, weight);
+        double index = service.calculate(height, weight);
 
         System.out.println("Индекс массы тела = " + Math.round(index));
     }
